requirements.txt
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
langchain-core==0.3.19
langchain-google-genai==2.0.5
langsmith==0.1.143
python-dotenv==1.0.1
streamlit==1.40.1
streamlit-confetti==0.1.0
mysql-connector==2.2.9
pymysql==1.1.1